start testing on one table
